I verified that the hipchat.py on toby can use room names, so do that.
Summary:
It's much easier to read (and figure out) than room numbers!

Publish to hipchat every time this script restarts.

If it's all the time, we want to know.  We also want to know, when
restarting the script, that it can actually successfully speak to
hipchat.

Test Plan:
Manually ran the hipchat function in notify.py, with the room_id as
'1s and 0s'.

None yet.  But once this is installed, I'll restart notify.py manually
on toby.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2003